feat: ini play mosh
