+++ components/molecules/src/index.ts
@@ -2,4 +2,3 @@ export * from "./ButtonGroup";
 export * from "./CircleIndicator";
 export * from "./GlowLogo";
 export * from "./Parallax";
-export * from "./ModalScreen";
