+++ components/organisms/src/ModalWindow/index.ts
@@ -1,0 +1,2 @@
+export * from "./ModalWindow";
+export * from "./styles";
