+++ components/organisms/src/ModalWindow/styles/index.ts
@@ -1,0 +1,1 @@
+export * from "./MoshModalWindow";
