+++ components/organisms/src/index.ts
@@ -2,5 +2,6 @@ export * from "./ButtonCTA";
 export * from "./ButtonSection";
 export * from "./FaqSection";
 export * from "./FooterLogo";
+export * from "./ModalWindow";
 export * from "./CountdownClock";
 export * from "./VaporGrid";
