+++ e2e/play-mosh-lol-e2e/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ["plugin:cypress/recommended", "../../.eslintrc.js"],
+};
