+++ e2e/play-mosh-lol-e2e/.local.env
@@ -1,0 +1,1 @@
+NEXT_TRANSLATE_PATH=apps/play-mosh-lol
