+++ e2e/play-mosh-lol-e2e/CHANGELOG.md
@@ -1,0 +1,40 @@
+## @sentrei/play-mosh-lol-e2e [1.0.1](https://github.com/sentrei/sentrei/compare/@sentrei/play-mosh-lol-e2e@1.0.0...@sentrei/play-mosh-lol-e2e@1.0.1) (2021-09-23)
+
+### Performance Improvements
+
+- ini tsc e2d ([299f23e](https://github.com/sentrei/sentrei/commit/299f23e4bc09c199ec375ac894f3e8d6709a94be))
+
+# @sentrei/play-mosh-lol-e2e 1.0.0 (2021-09-23)
+
+### Features
+
+- ini app mosh lol with json sort ([3bd1255](https://github.com/sentrei/sentrei/commit/3bd12550f6f1a2be250c0497c665e79e9d1ecd88))
+
+# @sentrei/app-sentrei-com-e2e 1.0.0 (2021-09-15)
+
+### Performance Improvements
+
+- ini app sentrei com ([13829af](https://github.com/sentrei/sentrei/commit/13829af431926ee2e59cb7860110ea2f80ef3148))
+- ini og sentrei com ([fe4ae7f](https://github.com/sentrei/sentrei/commit/fe4ae7f7cb4607645b7bea13f04b9fcbf88ddcd6))
+
+# @sentrei/app-e2e [1.1.0](https://github.com/sentrei/sentrei/compare/@sentrei/app-e2e@1.0.1...@sentrei/app-e2e@1.1.0) (2021-07-26)
+
+### Features
+
+- nx workspace revamp ([15dda56](https://github.com/sentrei/sentrei/commit/15dda56c923c7def734ddc4fe9411188c0366c1a))
+
+## @sentrei/app-e2e [1.0.1](https://github.com/sentrei/sentrei/compare/@sentrei/app-e2e@1.0.0...@sentrei/app-e2e@1.0.1) (2021-07-24)
+
+### Bug Fixes
+
+- e2e bug basePath ([29db229](https://github.com/sentrei/sentrei/commit/29db22962a5a894ad580b0907cb20e9d447f7c4d))
+
+# @sentrei/app-e2e 1.0.0 (2021-07-23)
+
+### Features
+
+- complete e2e packages ([a6fea5c](https://github.com/sentrei/sentrei/commit/a6fea5c42aad5f5e8c2c4df2b0fdfeedf7f7d914))
+
+### Performance Improvements
+
+- ini e2e testing for apps ([f81960c](https://github.com/sentrei/sentrei/commit/f81960c517085dad86e355461e61387bca504f67))
