+++ e2e/play-mosh-lol-e2e/cypress.json
@@ -1,0 +1,12 @@
+{
+  "chromeWebSecurity": false,
+  "fileServerFolder": ".",
+  "fixturesFolder": "./src/fixtures",
+  "integrationFolder": "./src/integration",
+  "modifyObstructiveCode": false,
+  "pluginsFile": "./src/plugins/index",
+  "screenshotsFolder": "../../dist/e2e/play-mosh-lol-e2e/screenshots",
+  "supportFile": "./src/support/index.ts",
+  "video": true,
+  "videosFolder": "../../dist/e2e/play-mosh-lol-e2e/videos"
+}
