+++ e2e/play-mosh-lol-e2e/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@sentrei/play-mosh-lol-e2e",
+  "version": "1.0.1",
+  "private": false,
+  "homepage": "https://github.com/sentrei/sentrei",
+  "bugs": {
+    "url": "https://github.com/sentrei/sentrei/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/sentrei/sentrei.git"
+  },
+  "license": "MIT",
+  "author": "Sentrei, Inc.",
+  "scripts": {
+    "fix": "cd ../.. && yarn run fix"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org"
+  }
+}
