+++ e2e/play-mosh-lol-e2e/project.json
@@ -1,0 +1,26 @@
+{
+  "projectType": "application",
+  "root": "e2e/play-mosh-lol-e2e",
+  "sourceRoot": "e2e/play-mosh-lol-e2e/src",
+  "targets": {
+    "e2e": {
+      "configurations": {
+        "production": {
+          "devServerTarget": "play-mosh-lol:serve:production"
+        }
+      },
+      "executor": "@nrwl/cypress:cypress",
+      "options": {
+        "cypressConfig": "e2e/play-mosh-lol-e2e/cypress.json",
+        "devServerTarget": "play-mosh-lol:serve",
+        "tsConfig": "e2e/play-mosh-lol-e2e/tsconfig.e2e.json"
+      }
+    },
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "options": {
+        "lintFilePatterns": ["e2e/play-mosh-lol-e2e/**/*.{js,ts}"]
+      }
+    }
+  }
+}
