+++ e2e/play-mosh-lol-e2e/src/fixtures/example.json
@@ -1,0 +1,1 @@
+{}
