+++ e2e/play-mosh-lol-e2e/src/integration/index.spec.ts
@@ -1,0 +1,7 @@
+export {};
+
+describe("Index", () => {
+  it("Index page should render correctly", () => {
+    cy.visit("/");
+  });
+});
