+++ e2e/play-mosh-lol-e2e/src/plugins/index.js
@@ -1,0 +1,5 @@
+const { preprocessTypescript } = require("@nrwl/cypress/plugins/preprocessor");
+
+module.exports = (on, config) => {
+  on("file:preprocessor", preprocessTypescript(config));
+};
