+++ e2e/play-mosh-lol-e2e/src/support/index.ts
@@ -1,0 +1,1 @@
+export {};
