+++ e2e/play-mosh-lol-e2e/tsconfig.e2e.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "outDir": "../../dist/e2e/play-mosh-lol-e2e/out",
+    "sourceMap": false,
+    "types": ["cypress", "node"]
+  },
+  "extends": "./tsconfig.json",
+  "include": ["src/**/*.ts", "src/**/*.js"]
+}
