--- libs/hooks/src/useModalScreen/index.ts
@@ -1,1 +1,0 @@
-export { modalScreenAtom, useModalScreen } from "./useModalScreen";
