+++ libs/hooks/src/useModalWindow/index.ts
@@ -1,0 +1,1 @@
+export { modalWindowAtom, useModalWindow } from "./useModalWindow";
